refactor: SingleOperation의 역할을 ArithmeticOperator에서 하도록 변경
