+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/ArithmeticOperator.kt
@@ -10,5 +10,14 @@ enum class ArithmeticOperator(val operation: String) {
         fun convertToArithmeticOperation(operation: String): ArithmeticOperator {
             return values().find { it.operation == operation } ?: throw IllegalArgumentException()
         }
+
+        fun ArithmeticOperator.calculate(firstNumber: Double, secondNumber: Double): Double {
+            return when (this) {
+                PLUS -> firstNumber + secondNumber
+                MINUS -> firstNumber - secondNumber
+                MULTIPLICATION -> firstNumber * secondNumber
+                DIVISION -> firstNumber / secondNumber
+            }
+        }
     }
 }
