--- domain/src/main/java/edu/nextstep/camp/calculator/domain/SingleOperation.kt
@@ -1,13 +1,0 @@
-package edu.nextstep.camp.calculator.domain
-
-object SingleOperation {
-
-    fun calculate(firstNumber: Double, secondNumber: Double, operator: ArithmeticOperator): Double {
-        return when (operator) {
-            ArithmeticOperator.PLUS -> firstNumber + secondNumber
-            ArithmeticOperator.MINUS -> firstNumber - secondNumber
-            ArithmeticOperator.MULTIPLICATION -> firstNumber * secondNumber
-            ArithmeticOperator.DIVISION -> firstNumber / secondNumber
-        }
-    }
-}
